the final countdown

+++ src/CajaAhorroPesos.java
@@ -7,7 +7,7 @@ public class CajaAhorroPesos extends Pesos{
     }
 
 	@Override
-	public void extraer(int monto) {
+	public void extraer(double monto) {
 		if(verificarMonto(monto) && verificarSaldo(monto) && monto/100 == 0) {
 			saldo-= monto;
 		
@@ -28,8 +28,8 @@ public class CajaAhorroPesos extends Pesos{
 
 	@Override
 	public void comprarUSD(Cliente cliente,double monto) {
-		int sinImpuestos = monto*70;
-		int conImpuestos = sinImpuestos+= sinImpuestos*0.3;
+		double sinImpuestos = monto*70;
+		double conImpuestos = sinImpuestos+= sinImpuestos*0.3;
 		if(verificarMonto(monto) && verificarSaldo(monto)) {
 			if(cliente.getUSD()!= null) {
 				saldo-= conImpuestos;
@@ -43,7 +43,7 @@ public class CajaAhorroPesos extends Pesos{
 	}
 
 	@Override
-	public void depositar(int monto) {
+	public void depositar(double monto) {
 		if(verificarMonto(monto) && monto/100 == 0) {
 			saldo+= monto;
