+++ src/TransaccionPesos.java
@@ -1,10 +1,10 @@
 public abstract interface TransaccionPesos {      
 
 	
-	abstract void extraer(int monto);
-	abstract void transferir(Cliente cliente, int monto);
-	abstract void comprarUSD(Cliente cliente, int monto);
-	abstract void depositar(int monto);
+	abstract void extraer(double monto);
+	abstract void transferir(Cliente cliente, double monto);
+	abstract void comprarUSD(Cliente cliente, double monto);
+	abstract void depositar(double monto);
